// src/app/shares/models/merchOrder.model.ts
export interface order {
    id: Number;
    name: String;
    ustEmail: String;
    year: String;
    section: String;
    studentNo: String;
    merchName: String;
    quantity: Number;
    contactNo: number;
    address: String;
    modeOfPayment: String;
    whatMerchNext: String;
    price: number;
}